requirements.txt
alembic==1.6.2
Authlib==0.15.3
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
cryptography==3.4.7
decorator==5.0.9
dnspython==2.1.0
email-validator==1.1.2
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
Mako==1.1.4
MarkupSafe==2.0.0
pbr==5.6.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.15
sqlalchemy-migrate==0.13.0
sqlparse==0.4.1
Tempita==0.5.2
urllib3==1.26.4
Werkzeug==2.0.0
WTForms==2.3.3